BACKEND+FRONTEND: deleting of users (and ALL his/her data) through moderation panel

+++ src/frontend/dev_puzzles/moderation/core/user-moderation/user-moderation.component.ts
@@ -22,7 +22,7 @@ export class UserModerationComponent implements OnInit {
 
 
   deleteUser():void{
-    if(window.confirm("Are you sure you want to delete user " + this.user.name + '?')){
+    if(window.confirm("Are you sure you want to delete user " + this.user.name + ' and all of its data?')){
       this.moderationService.deleteUser(this.user._id).subscribe(this.userDeleted.bind(this));
     }
   }
